Changed Long to Integer in Client files

+++ src/main/java/com/tp_note/tp_note/data/repository/ClientRepository.java
@@ -19,17 +19,17 @@ public class ClientRepository {
 		return this.jdbcTemplate.queryForList(sql);
 	}
 
-	public List getContrats(Long clientId) {
+	public List getContrats(Integer clientId) {
 		String sql = "SELECT * FROM contrat WHERE id_client="+clientId.toString()+";";
 		return this.jdbcTemplate.queryForList(sql);
 	}
 
-	public List findContratByIdAndContratId(Long clientId, Long contratId) {
+	public List findContratByIdAndContratId(Integer clientId, Integer contratId) {
 		String sql = "SELECT * FROM contrat WHERE id_client="+clientId.toString()+" AND id_contrat="+contratId.toString()+";";
 		return this.jdbcTemplate.queryForList(sql);
 	}
 
-	public List getClientsFromConseiller(Long conseillerId) {
+	public List getClientsFromConseiller(Integer conseillerId) {
 		String sql = "SELECT * FROM client WHERE id_conseiller="+conseillerId.toString()+";";
 		return this.jdbcTemplate.queryForList(sql);
 	}
